// backend/src/config/tmdb.d.ts
export declare const TMDB_CONFIG: {
    apiKey: string;
    baseUrl: string;
    imageUrl: string;
    language: string;
    mockMode: boolean;
};
export declare const IMAGE_SIZES: {
    poster: {
        small: string;
        medium: string;
        large: string;
        original: string;
    };
    backdrop: {
        small: string;
        medium: string;
        large: string;
        original: string;
    };
};
export declare function getImageUrl(path: string, size?: string): string;
//# sourceMappingURL=tmdb.d.ts.map